// src/app/modules/admin/models/ship-to-locations.model.ts
export class ShipToLocations {
  locationId: number=0;
  clientId: number=0;
  orgId: string='';
  invOrgId: string='';
  locationName: string='';
  address1: string='';
  address2: string='';
  address3: string='';
  address4: string='';
  city: string='';
  stateProvinceCode: string='';
  postalCode: string='';
  countryCode: string='';
  companyName: string='';
  contactName: string='';
  phoneNumber: string='';
  emailId: string='';
}